code fixing after review, further implementation of richfaces. The spice must flow.

+++ BookCatalog/BookCatalog-cdi/src/main/java/ua/softserve/bandr/persistence/facade/AbstractFacadeInt.java
@@ -18,5 +18,5 @@ public interface AbstractFacadeInt<T extends Persistable> {
 
 	List<T> getPagedFilteredSorted(Integer startWith, Integer pageSize, Map<String, String> filterText, Map<String, Boolean> sortingOrder);
 
-	Integer getRecordCount(Map<String, String> filter);
+	Long getRecordCount(Map<String, String> filter);
 }
